Move all task related components to specific components package
