+++ app/src/main/java/nl/tue/san/sanseminar/components/tasks/TaskSetIO.java
@@ -1,4 +1,4 @@
-package nl.tue.san.sanseminar.components;
+package nl.tue.san.sanseminar.components.tasks;
 
 import org.json.JSONArray;
 import org.json.JSONException;
